Remove usage of Network.post(). Will look into cleaning up the makeRequestWithSideEffects next
