+++ src/libs/actions/App.js
@@ -286,6 +286,18 @@ function openProfile() {
     Navigation.navigate(ROUTES.SETTINGS_PROFILE);
 }
 
+function getShortLivedAuthToken() {
+    // eslint-disable-next-line rulesdir/no-api-side-effects-method
+    return API.makeRequestWithSideEffects('OpenOldDotLink')
+        .then((response) => {
+            if (response && response.shortLivedAuthToken) {
+                return Promise.resolve(response.shortLivedAuthToken);
+            }
+
+            return Promise.reject();
+        });
+}
+
 export {
     setLocale,
     setLocaleAndNavigate,
@@ -294,4 +306,5 @@ export {
     openProfile,
     openApp,
     reconnectApp,
+    getShortLivedAuthToken,
 };
